requirements.txt
autopep8==2.0.2
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
pycodestyle==2.10.0
python-dotenv==1.0.0
SQLAlchemy==1.4.18
typing_extensions==4.5.0
Werkzeug==2.2.3